+++ app/internal/modules/catalog/category/usecase/admin-api/category.go
@@ -101,7 +101,9 @@ func (uc *Category) Store(ctx context.Context, item entity.Category) error {
 		return mrcore.FactoryErrUseCaseEntityVersionInvalid.New()
 	}
 
-	if err := uc.storage.IsExists(ctx, item.ID); err != nil {
+	// предварительная проверка существования записи нужна для того,
+	// чтобы при Update быть уверенным, что отсутствие записи из-за VersionInvalid
+	if _, err := uc.storage.FetchStatus(ctx, item.ID); err != nil {
 		return uc.usecaseHelper.WrapErrorEntityNotFoundOrFailed(err, entity.ModelNameCategory, item.ID)
 	}
 
@@ -129,7 +131,7 @@ func (uc *Category) ChangeStatus(ctx context.Context, item entity.Category) erro
 		return mrcore.FactoryErrUseCaseEntityVersionInvalid.New()
 	}
 
-	currentStatus, err := uc.storage.FetchStatus(ctx, item)
+	currentStatus, err := uc.storage.FetchStatus(ctx, item.ID)
 
 	if err != nil {
 		return uc.usecaseHelper.WrapErrorEntityNotFoundOrFailed(err, entity.ModelNameCategory, item.ID)