+++ app/internal/modules/catalog/category/usecase/api/category_interface.go
@@ -4,10 +4,11 @@ import (
 	"context"
 
 	"github.com/google/uuid"
+	"github.com/mondegor/go-webcore/mrenum"
 )
 
 type (
 	CategoryStorage interface {
-		IsExists(ctx context.Context, rowID uuid.UUID) error
+		FetchStatus(ctx context.Context, rowID uuid.UUID) (mrenum.ItemStatus, error)
 	}
 )