+++ app/internal/modules/catalog/product/infrastructure/repository/admin-api/product_postgres.go
@@ -220,7 +220,9 @@ func (re *ProductPostgres) FetchIdByArticle(ctx context.Context, article string)
 	return rowID, err
 }
 
-func (re *ProductPostgres) FetchStatus(ctx context.Context, row entity.Product) (mrenum.ItemStatus, error) {
+// FetchStatus
+// result: mrenum.ItemStatus - exists, ErrStorageNoRowFound - not exists, error - query error
+func (re *ProductPostgres) FetchStatus(ctx context.Context, rowID mrtype.KeyInt32) (mrenum.ItemStatus, error) {
 	sql := `
 		SELECT
 			product_status
@@ -235,35 +237,13 @@ func (re *ProductPostgres) FetchStatus(ctx context.Context, row entity.Product) 
 	err := re.client.QueryRow(
 		ctx,
 		sql,
-		row.ID,
+		rowID,
 		mrenum.ItemStatusRemoved,
 	).Scan(
 		&status,
 	)
 
 	return status, err
-}
-
-// IsExists
-// result: nil - exists, ErrStorageNoRowFound - not exists, error - query error
-func (re *ProductPostgres) IsExists(ctx context.Context, rowID mrtype.KeyInt32) error {
-	sql := `
-		SELECT
-			product_id
-		FROM
-			` + module.DBSchema + `.products
-		WHERE
-			product_id = $1 AND product_status <> $2
-		LIMIT 1;`
-
-	return re.client.QueryRow(
-		ctx,
-		sql,
-		rowID,
-		mrenum.ItemStatusRemoved,
-	).Scan(
-		&rowID,
-	)
 }
 
 func (re *ProductPostgres) Insert(ctx context.Context, row entity.Product) (mrtype.KeyInt32, error) {
@@ -373,8 +353,8 @@ func (re *ProductPostgres) Delete(ctx context.Context, rowID mrtype.KeyInt32) er
 			` + module.DBSchema + `.products
 		SET
 			tag_version = tag_version + 1,
-			updated_at = NOW(),
 			product_article = NULL,
+			updated_at = NOW(),
 			order_index = NULL,
 			product_status = $2
 		WHERE