+++ app/internal/modules/catalog/trademark/usecase/admin-api/trademark.go
@@ -90,7 +90,9 @@ func (uc *Trademark) Store(ctx context.Context, item entity.Trademark) error {
 		return mrcore.FactoryErrUseCaseEntityVersionInvalid.New()
 	}
 
-	if err := uc.storage.IsExists(ctx, item.ID); err != nil {
+	// предварительная проверка существования записи нужна для того,
+	// чтобы при Update быть уверенным, что отсутствие записи из-за VersionInvalid
+	if _, err := uc.storage.FetchStatus(ctx, item.ID); err != nil {
 		return uc.usecaseHelper.WrapErrorEntityNotFoundOrFailed(err, entity.ModelNameTrademark, item.ID)
 	}
 
@@ -118,7 +120,7 @@ func (uc *Trademark) ChangeStatus(ctx context.Context, item entity.Trademark) er
 		return mrcore.FactoryErrUseCaseEntityVersionInvalid.New()
 	}
 
-	currentStatus, err := uc.storage.FetchStatus(ctx, item)
+	currentStatus, err := uc.storage.FetchStatus(ctx, item.ID)
 
 	if err != nil {
 		return uc.usecaseHelper.WrapErrorEntityNotFoundOrFailed(err, entity.ModelNameTrademark, item.ID)