+++ app/pkg/modules/catalog/category_api.go
@@ -9,7 +9,11 @@ import (
 
 type (
 	CategoryAPI interface {
-		// CheckingAvailability - error: FactoryErrCategoryRequired | FactoryErrCategoryNotFound | Failed
+		// CheckingAvailability - error:
+		//   - FactoryErrCategoryRequired
+		//   - FactoryErrCategoryNotAvailable
+		//   - FactoryErrCategoryNotFound
+		//   - Failed
 		CheckingAvailability(ctx context.Context, itemID uuid.UUID) error
 	}
 )
@@ -17,6 +21,9 @@ type (
 var (
 	FactoryErrCategoryRequired = mrerr.NewFactory(
 		"errCatalogCategoryRequired", mrerr.ErrorKindUser, "category ID is required")
+
+	FactoryErrCategoryNotAvailable = mrerr.NewFactory(
+		"errCatalogCategoryNotAvailable", mrerr.ErrorKindUser, "category with ID={{ .id }} is not available")
 
 	FactoryErrCategoryNotFound = mrerr.NewFactory(
 		"errCatalogCategoryNotFound", mrerr.ErrorKindUser, "category with ID={{ .id }} not found")