+++ app/pkg/modules/catalog/trademark_api.go
@@ -9,7 +9,11 @@ import (
 
 type (
 	TrademarkAPI interface {
-		// CheckingAvailability - error: FactoryErrTrademarkRequired | FactoryErrTrademarkNotFound |
+		// CheckingAvailability - error:
+		//    - FactoryErrTrademarkRequired
+		//	  - FactoryErrTrademarkNotAvailable
+		//	  - FactoryErrTrademarkNotFound
+		//	  - Failed
 		CheckingAvailability(ctx context.Context, itemID mrtype.KeyInt32) error
 	}
 )
@@ -17,6 +21,9 @@ type (
 var (
 	FactoryErrTrademarkRequired = mrerr.NewFactory(
 		"errCatalogTrademarkRequired", mrerr.ErrorKindUser, "trademark ID is required")
+
+	FactoryErrTrademarkNotAvailable = mrerr.NewFactory(
+		"errCatalogTrademarkNotAvailable", mrerr.ErrorKindUser, "trademark with ID={{ .id }} is not available")
 
 	FactoryErrTrademarkNotFound = mrerr.NewFactory(
 		"errCatalogTrademarkNotFound", mrerr.ErrorKindUser, "trademark with ID={{ .id }} not found")